requirements.txt
bs4
selenium
python-docx